backend/requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.2.0
django-stubs==4.2.3
django-stubs-ext==4.2.2
djangorestframework==3.14.0
djangorestframework-stubs==3.14.2
flake8==6.1.0
googletrans==3.1.0a0
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
types-pytz==2023.3.0.1
types-PyYAML==6.0.12.11
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing_extensions==4.7.1
urllib3==2.0.4
